Move the Recent Projects and Recent Files widgets into the respective modules.
Remove the need for friend access from IDE dashboard and allow platform
applications to use these widgets in dashboard configurations.
