+++ ide/projectui/nbproject/project.properties
@@ -16,7 +16,7 @@
 # under the License.
 
 javac.compilerargs=-Xlint:unchecked
-javac.source=1.8
+javac.release=17
 spec.version.base=1.84.0
 
 javadoc.arch=${basedir}/arch.xml
