+++ ide/utilities/nbproject/project.properties
@@ -17,7 +17,7 @@
 
 #javac.compilerargs=-Xlint:unchecked
 
-javac.source=1.8
+javac.release=17
 javadoc.arch=${basedir}/arch.xml
 
 test.config.stableBTD.includes=**/*Test.class
