+++ nb/ide.dashboard/src/org/netbeans/modules/ide/dashboard/resources/layer.xml
@@ -22,54 +22,48 @@
 <!DOCTYPE filesystem PUBLIC "-//NetBeans//DTD Filesystem 1.2//EN" "http://www.netbeans.org/dtds/filesystem-1_2.dtd">
 <filesystem>
 
-  <folder name="Dashboard">
-    <folder name="Main">
-      <attr name="showOnStartup" boolvalue="false"/>
-      <file name="NetBeansWidget.shadow">
-        <attr name="originalFile" stringvalue="Dashboard/Widgets/NetBeansWidget.instance"/>
-        <attr name="position" intvalue="100"/>
-      </file>
-      <file name="RecentProjectsWidget.shadow">
-        <attr name="originalFile" stringvalue="Dashboard/Widgets/RecentProjectsWidget.instance"/>
-        <attr name="position" intvalue="200"/>
-      </file>
-      <file name="RecentFilesWidget.shadow">
-        <attr name="originalFile" stringvalue="Dashboard/Widgets/RecentFilesWidget.instance"/>
-        <attr name="position" intvalue="300"/>
-      </file>
-      <file name="PluginsWidget.shadow">
-        <attr name="originalFile" stringvalue="Dashboard/Widgets/PluginsWidget.instance"/>
-        <attr name="position" intvalue="400"/>
-      </file>
-      <file name="HelpWidget.shadow">
-        <attr name="originalFile" stringvalue="Dashboard/Widgets/HelpWidget.instance"/>
-        <attr name="position" intvalue="500"/>
-      </file>
-      <file name="AppearanceWidget.shadow">
-        <attr name="originalFile" stringvalue="Dashboard/Widgets/AppearanceWidget.instance"/>
-        <attr name="position" intvalue="600"/>
-      </file>
+    <folder name="Dashboard">
+        <folder name="Main">
+            <attr name="showOnStartup" boolvalue="false"/>
+            <file name="NetBeansWidget.shadow">
+                <attr name="originalFile" stringvalue="Dashboard/Widgets/NetBeansWidget.instance"/>
+                <attr name="position" intvalue="100"/>
+            </file>
+            <file name="RecentProjectsWidget.shadow">
+                <attr name="originalFile" stringvalue="Dashboard/Widgets/RecentProjectsWidget.instance"/>
+                <attr name="position" intvalue="200"/>
+            </file>
+            <file name="RecentFilesWidget.shadow">
+                <attr name="originalFile" stringvalue="Dashboard/Widgets/RecentFilesWidget.instance"/>
+                <attr name="position" intvalue="300"/>
+            </file>
+            <file name="PluginsWidget.shadow">
+                <attr name="originalFile" stringvalue="Dashboard/Widgets/PluginsWidget.instance"/>
+                <attr name="position" intvalue="400"/>
+            </file>
+            <file name="HelpWidget.shadow">
+                <attr name="originalFile" stringvalue="Dashboard/Widgets/HelpWidget.instance"/>
+                <attr name="position" intvalue="500"/>
+            </file>
+            <file name="AppearanceWidget.shadow">
+                <attr name="originalFile" stringvalue="Dashboard/Widgets/AppearanceWidget.instance"/>
+                <attr name="position" intvalue="600"/>
+            </file>
+        </folder>
+        <folder name="Widgets">
+            <file name="AppearanceWidget.instance">
+                <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.AppearanceWidget"/>
+            </file>
+            <file name="HelpWidget.instance">
+                <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.HelpWidget"/>
+            </file>
+            <file name="NetBeansWidget.instance">
+                <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.NetBeansWidget"/>
+            </file>
+            <file name="PluginsWidget.instance">
+                <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.PluginsWidget"/>
+            </file>
+        </folder>
     </folder>
-    <folder name="Widgets">
-      <file name="AppearanceWidget.instance">
-        <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.AppearanceWidget"/>
-      </file>
-      <file name="HelpWidget.instance">
-        <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.HelpWidget"/>
-      </file>
-      <file name="NetBeansWidget.instance">
-        <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.NetBeansWidget"/>
-      </file>
-      <file name="PluginsWidget.instance">
-        <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.PluginsWidget"/>
-      </file>
-      <file name="RecentFilesWidget.instance">
-        <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.RecentFilesWidget"/>
-      </file>
-      <file name="RecentProjectsWidget.instance">
-        <attr name="instanceClass" stringvalue="org.netbeans.modules.ide.dashboard.RecentProjectsWidget"/>
-      </file>
-    </folder>
-  </folder>
   
 </filesystem>
